Fully working version

+++ includes/signup.php
@@ -10,23 +10,26 @@
     $password_confirm = $_POST['password_confirm'];
 
     if($password === $password_confirm) {
-
-        if($_FILES['avatar']['name']) {
-            $path = 'uploads/' . time() . $_FILES['avatar']['name'];
-            if(!move_uploaded_file($_FILES['avatar']['tmp_name'], '../' . $path)) {
-                $_SESSION['message'] = 'Ошибка при загрузке изображения';
-                header('Location: ../reg.php');
+        if($full_name != NULL && $login != NULL && $password != NULL && $password_confirm != NULL) {
+            if ($_FILES['avatar']['name']) {
+                $path = 'uploads/' . time() . $_FILES['avatar']['name'];
+                if (!move_uploaded_file($_FILES['avatar']['tmp_name'], '../' . $path)) {
+                    $_SESSION['message'] = 'Ошибка при загрузке изображения';
+                    header('Location: ../reg.php');
+                }
+            } else {
+                $path = NULL;
             }
-        }
-        else{
-            $path = NULL;
-        }
 
-        $password = md5($password);
+            $password = md5($password);
 
-        mysqli_query($connect, "INSERT INTO `users` (`id`, `full_name`, `login`, `email`, `avatar`, `password`) VALUES (NULL, '$full_name', '$login', '$email', '$path', '$password')");
-        $_SESSION['message'] = 'Регистрация прошла успешно!';
-        header('Location: ../auth.php');
+            mysqli_query($connect, "INSERT INTO `users` (`id`, `full_name`, `login`, `email`, `avatar`, `password`) VALUES (NULL, '$full_name', '$login', '$email', '$path', '$password')");
+            $_SESSION['message'] = 'Регистрация прошла успешно!';
+            header('Location: ../auth.php');
+        } else {
+            $_SESSION['message'] = 'Заполните все обязательные поля';
+            header('Location: ../reg.php');
+        }
 
     } else {
         $_SESSION['message'] = 'Пароли не совпадают';
